Refactored

+++ BookDescription/ISBN.swift
@@ -76,7 +76,7 @@ struct ISBN : CustomStringConvertible, Equatable {
             let check = ISBN.reckonISBN13CheckDigit(prelim)
             self.digits = 10 * prelim + UInt64(check)
         } else {
-            let parsed = UInt64(ISBN.removeDashes(number))
+            let parsed = ISBN.removeDashes(number)
             if parsed < 9780000000000 {
                 let prelim = 978000000000 + (parsed / 10)
                 let check = ISBN.reckonISBN13CheckDigit(prelim)
